refactor: extract estimated call components
